+++ lib_comfyui/comfyui_process.py
@@ -4,7 +4,7 @@ import os
 import signal
 import subprocess
 import sys
-from lib_comfyui import ipc, torch_utils, argv_conversion, parallel_utils
+from lib_comfyui import ipc, torch_utils, argv_conversion, ipc_callback
 from lib_comfyui.webui import settings
 from lib_comfyui.comfyui import pre_main
 
@@ -24,8 +24,8 @@ def start():
         print('[sd-webui-comfyui]', f'could not find ComfyUI under directory "{install_location}". The server will NOT be started.', file=sys.stderr)
         return
 
-    ipc.current_callback_listeners = {'webui': parallel_utils.CallbackWatcher(ipc.call_fully_qualified, 'webui')}
-    ipc.current_callback_proxies = {'comfyui': parallel_utils.CallbackProxy('comfyui')}
+    ipc.current_callback_listeners = {'webui': ipc_callback.CallbackWatcher(ipc.call_fully_qualified, 'webui')}
+    ipc.current_callback_proxies = {'comfyui': ipc_callback.CallbackProxy('comfyui')}
     ipc.start_callback_listeners()
     atexit.register(stop)
     start_comfyui_process(install_location)
