+++ lib_comfyui/ipc_payload.py
@@ -2,92 +2,19 @@ import contextlib
 import dataclasses
 import os
 import pickle
+import portalocker
 import tempfile
-import threading
 import time
 import logging
 from multiprocessing.shared_memory import SharedMemory
 from pathlib import Path
 from typing import Optional, Any, IO, Union
-import portalocker
-
-
-class CallbackWatcher:
-    def __init__(self, callback, name: str):
-        self._callback = callback
-        self._res_sender = IpcSender(f'res_{name}')
-        self._args_receiver = IpcReceiver(f'args_{name}')
-        self._producer_thread = None
-
-    def start(self):
-        def thread_loop():
-            while self._producer_thread.is_running():
-                self.attend_consumer(self._callback, timeout=0.5)
-
-        self._producer_thread = StoppableThread(target=thread_loop, daemon=True)
-        self._producer_thread.start()
-
-    def stop(self):
-        if self._producer_thread is None:
-            return
-
-        self._producer_thread.join()
-        self._producer_thread = None
-
-    def is_running(self):
-        return self._producer_thread and self._producer_thread.is_running()
-
-    def attend_consumer(self, callback, timeout: float = None):
-        try:
-            args, kwargs = self._args_receiver.recv(timeout=timeout)
-        except TimeoutError:
-            return
-
-        try:
-            self._res_sender.send(callback(*args, **kwargs))
-        except Exception as e:
-            self._res_sender.send(RemoteError(e))
-
-
-class CallbackProxy:
-    def __init__(self, name):
-        self._res_receiver = IpcReceiver(f'res_{name}')
-        self._args_sender = IpcSender(f'args_{name}')
-
-    def get(self, args=None, kwargs=None):
-        self._args_sender.send((args if args is not None else (), kwargs if kwargs is not None else {}))
-        res = self._res_receiver.recv()
-        if isinstance(res, RemoteError):
-            raise res.error from res
-        else:
-            return res
-
-
-class RemoteError(Exception):
-    def __init__(self, error):
-        self.error = error
-
-
-class StoppableThread(threading.Thread):
-    def __init__(self, *args, **kwargs):
-        super().__init__(*args, **kwargs)
-        self._stop_event = threading.Event()
-
-    def stop(self):
-        self._stop_event.set()
-
-    def join(self, *args, **kwargs) -> None:
-        self.stop()
-        super().join(*args, **kwargs)
-
-    def is_running(self):
-        return not self._stop_event.is_set()
 
 
 class IpcPayload:
-    def __init__(self, name, strategy=None, lock_directory=None):
+    def __init__(self, name, strategy, lock_directory=None):
         self._name = name
-        self._strategy = strategy if strategy is not None else OsIpcStrategy(f'ipc_payload_{name}')
+        self._strategy = strategy
         lock_directory = tempfile.gettempdir() if lock_directory is None else lock_directory
         self._lock_path = Path(lock_directory, f'ipc_payload_{name}')
         with self.get_lock() as lock_file:
