+++ requirements.txt
@@ -1,1 +1,1 @@
-watchdog>=3.0.0
+portalocker>=2.7.0
